Refs #32880 - move rails override to foreman_overrides.js
